# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Inovare TI API",
    "version": "1.0.0",
    "description": "OpenAPI specification gerada automaticamente — resumo dos endpoints do backend Inovare TI."
  },
  "servers": [
    {
      "url": "http://localhost:8085"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
...
    "/financeiro/contaazul/callback": {
      "get": {
        "summary": "Callback OAuth2 ContaAzul (redirect para frontend)",
        "parameters": [ { "name": "code", "in": "query", "required": true, "schema": { "type": "string" } } ],
        "responses": { "302": { "description": "Redirect para frontend com sucesso" } }
      }
    },
    "/financeiro/contaazul/force-refresh": {
      "post": {
        "summary": "Forçar refresh do token ContaAzul (ADMIN)",
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": {
            "description": "Token renovado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authorized": { "type": "boolean" },
                    "expiresAt": { "type": "string", "format": "date-time" },
                    "refreshedAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
          },
          "500": {
            "description": "Erro interno",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } }
          }
        }
      }
    },
...
  },
...
}